+++ src/test/java/org/opensearch/neuralsearch/query/HybridQueryFilterIT.java
@@ -113,7 +113,7 @@ public class HybridQueryFilterIT extends BaseNeuralSearchIT {
                 Collections.singletonList(new KNNFieldConfig(TEST_KNN_VECTOR_FIELD_NAME_1, TEST_DIMENSION, TEST_SPACE_TYPE))
             );
 
-            addKnnDoc(
+            indexTheDocument(
                 indexName,
                 "1",
                 Collections.singletonList(TEST_KNN_VECTOR_FIELD_NAME_1),
@@ -121,16 +121,19 @@ public class HybridQueryFilterIT extends BaseNeuralSearchIT {
                 Collections.singletonList(TEXT_FIELD_1_NAME),
                 Collections.singletonList(TEXT_FIELD_VALUE_2_DUNES),
                 List.of(),
-                List.of(),
+                Map.of(),
                 List.of(INTEGER_FIELD_1_STOCK),
                 List.of(INTEGER_FIELD_STOCK_1_25),
                 List.of(KEYWORD_FIELD_2_CATEGORY),
                 List.of(KEYWORD_FIELD_CATEGORY_1_DRAMA),
                 List.of(),
-                List.of()
+                List.of(),
+                List.of(),
+                List.of(),
+                null
             );
 
-            addKnnDoc(
+            indexTheDocument(
                 indexName,
                 "2",
                 Collections.singletonList(TEST_KNN_VECTOR_FIELD_NAME_1),
@@ -138,16 +141,19 @@ public class HybridQueryFilterIT extends BaseNeuralSearchIT {
                 Collections.singletonList(TEXT_FIELD_1_NAME),
                 Collections.singletonList(TEXT_FIELD_VALUE_1_DUNES),
                 List.of(),
-                List.of(),
+                Map.of(),
                 List.of(INTEGER_FIELD_1_STOCK),
                 List.of(INTEGER_FIELD_STOCK_2_22),
                 List.of(KEYWORD_FIELD_2_CATEGORY),
                 List.of(KEYWORD_FIELD_CATEGORY_1_DRAMA),
                 List.of(),
-                List.of()
+                List.of(),
+                List.of(),
+                List.of(),
+                null
             );
 
-            addKnnDoc(
+            indexTheDocument(
                 indexName,
                 "3",
                 Collections.singletonList(TEST_KNN_VECTOR_FIELD_NAME_1),
@@ -155,16 +161,19 @@ public class HybridQueryFilterIT extends BaseNeuralSearchIT {
                 Collections.singletonList(TEXT_FIELD_1_NAME),
                 Collections.singletonList(TEXT_FIELD_VALUE_3_MI_1),
                 List.of(),
-                List.of(),
+                Map.of(),
                 List.of(INTEGER_FIELD_1_STOCK),
                 List.of(INTEGER_FIELD_STOCK_3_256),
                 List.of(KEYWORD_FIELD_2_CATEGORY),
                 List.of(KEYWORD_FIELD_CATEGORY_2_ACTION),
                 List.of(),
-                List.of()
+                List.of(),
+                List.of(),
+                List.of(),
+                null
             );
 
-            addKnnDoc(
+            indexTheDocument(
                 indexName,
                 "4",
                 List.of(),
@@ -172,16 +181,19 @@ public class HybridQueryFilterIT extends BaseNeuralSearchIT {
                 Collections.singletonList(TEXT_FIELD_1_NAME),
                 Collections.singletonList(TEXT_FIELD_VALUE_4_MI_2),
                 List.of(),
-                List.of(),
+                Map.of(),
                 List.of(INTEGER_FIELD_1_STOCK),
                 List.of(INTEGER_FIELD_STOCK_4_25),
                 List.of(KEYWORD_FIELD_2_CATEGORY),
                 List.of(KEYWORD_FIELD_CATEGORY_2_ACTION),
                 List.of(),
-                List.of()
+                List.of(),
+                List.of(),
+                List.of(),
+                null
             );
 
-            addKnnDoc(
+            indexTheDocument(
                 indexName,
                 "5",
                 List.of(),
@@ -189,16 +201,19 @@ public class HybridQueryFilterIT extends BaseNeuralSearchIT {
                 Collections.singletonList(TEXT_FIELD_1_NAME),
                 Collections.singletonList(TEXT_FIELD_VALUE_5_TERMINAL),
                 List.of(),
-                List.of(),
+                Map.of(),
                 List.of(INTEGER_FIELD_1_STOCK),
                 List.of(INTEGER_FIELD_STOCK_5_20),
                 List.of(KEYWORD_FIELD_2_CATEGORY),
                 List.of(KEYWORD_FIELD_CATEGORY_1_DRAMA),
                 List.of(),
-                List.of()
+                List.of(),
+                List.of(),
+                List.of(),
+                null
             );
 
-            addKnnDoc(
+            indexTheDocument(
                 indexName,
                 "6",
                 List.of(),
@@ -206,13 +221,16 @@ public class HybridQueryFilterIT extends BaseNeuralSearchIT {
                 Collections.singletonList(TEXT_FIELD_1_NAME),
                 Collections.singletonList(TEXT_FIELD_VALUE_6_AVENGERS),
                 List.of(),
-                List.of(),
+                Map.of(),
                 List.of(INTEGER_FIELD_1_STOCK),
                 List.of(INTEGER_FIELD_STOCK_5_20),
                 List.of(KEYWORD_FIELD_2_CATEGORY),
                 List.of(KEYWORD_FIELD_CATEGORY_3_SCI_FI),
                 List.of(),
-                List.of()
+                List.of(),
+                List.of(),
+                List.of(),
+                null
             );
         }
     }